Add TechDataHandler and TechDataPatcher
